refactor image downloading from ValueConverter to ViewModel in order to asyncify
