--- src/NuGet.Clients/NuGet.PackageManagement.UI/Converters/IconUrlToVisibilityConverter.cs
@@ -1,32 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Diagnostics;
-using System.Globalization;
-using System.Windows;
-using System.Windows.Data;
-
-namespace NuGet.PackageManagement.UI
-{
-    public class IconUrlToVisibilityConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            var uri = value as Uri;
-            if (uri == null)
-            {
-                return Visibility.Collapsed;
-            }
-
-            return Visibility.Visible;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            // no op
-            Debug.Fail("Not Implemented");
-            return null;
-        }
-    }
-}
