+++ test/NuGet.Clients.Tests/NuGet.PackageManagement.UI.Test/Resources/testpackageicon.png
